# admin website/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Keto Diet Doctor Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: doctorLogin
      tags: [auth]
      summary: Doctor login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged in doctor
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    put:
      operationId: updateDoctorProfile
      tags: [auth]
      summary: Update doctor profile (name, email, designation, username)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/force-change:
    put:
      operationId: forceChangePassword
      tags: [auth]
      summary: Force password change (first login — no current password required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForceChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password:
    put:
      operationId: changeDoctorPassword
      tags: [auth]
      summary: Change doctor password (requires current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeDoctorPasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /foods:
    get:
      operationId: getFoods
      tags: [foods]
      summary: Get all foods
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of foods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Food"
    post:
      operationId: createFood
      tags: [foods]
      summary: Create a new food item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodRequest"
      responses:
        "201":
          description: Food created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
...
  /foods/{foodId}:
    get:
      operationId: getFood
      tags: [foods]
      summary: Get a single food item
      parameters:
        - name: foodId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Food item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
    put:
      operationId: updateFood
      tags: [foods]
      summary: Update a food item
      parameters:
        - name: foodId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFoodRequest"
      responses:
        "200":
          description: Food updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
    delete:
      operationId: deleteFood
      tags: [foods]
      summary: Delete a food item
      parameters:
        - name: foodId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Food deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        doctor:
          $ref: "#/components/schemas/Doctor"
        token:
          type: string
      required:
        - doctor
        - token
    Doctor:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
        designation:
          type: string
        role:
          type: string
          enum: [admin, moderator]
        mustChangePassword:
          type: boolean
      required:
        - id
        - username
        - name
        - email
        - role
    UpdateDoctorProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        designation:
          type: string
        username:
          type: string
          minLength: 3
          maxLength: 100
      required:
        - name
        - email
        - username
    ForceChangePasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 1
      required:
        - newPassword
        - confirmPassword
    ChangeDoctorPasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 1
      required:
        - currentPassword
        - newPassword
        - confirmPassword
...
    Food:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        calories:
          type: number
        imageUrl:
          type: string
        description:
          type: string
        indicator:
          type: string
          enum: [vegi, fruit, non-vegi, recipe]
        isActive:
          type: boolean
      required:
        - id
        - name
        - category
        - carbs
        - fat
        - protein
        - calories
        - isActive
    CreateFoodRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        calories:
          type: number
        imageUrl:
          type: string
        description:
          type: string
        indicator:
          type: string
          enum: [vegi, fruit, non-vegi, recipe]
      required:
        - name
        - category
        - carbs
        - fat
        - protein
        - calories
    UpdateFoodRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        carbs:
          type: number
        fat:
          type: number
        protein:
          type: number
        calories:
          type: number
        imageUrl:
          type: string
        description:
          type: string
        indicator:
          type: string
          enum: [vegi, fruit, non-vegi, recipe]
        isActive:
          type: boolean
...